Add files via upload
This project acquires questions from an API that contains tons of data from Jeopardy. It then gives the user an option of three techniques to acquire and use that data. Such as, Async/Await, XML callbacks, and lastly fetch and "then" chaining, each separated using Object Oriented Programming.

it includes a simple user interface and the theme song from Jeopardy as a nod to the show.